Add keyboard manager to MouseDrawnTool
