+++ client/paintapp.py
@@ -1,24 +1,12 @@
 import random
 import string
 
-import cv2
-from skimage.draw import line, disk, rectangle_perimeter
-from skimage.segmentation import flood
-import numba
-from numba import vectorize, uint8, boolean, int32
 import numpy as np
 from kivy.app import App
-from kivy.properties import ObjectProperty
-from kivy.uix.floatlayout import FloatLayout
-from kivy.uix.widget import Widget
-from numba import jit
 from kivy.core.window import Window
-
-from skimage.color import rgb2gray
+from kivy.uix.floatlayout import FloatLayout
 
 from client.screens.common import MouseDrawnTool
-from kivy.graphics.texture import Texture
-import time
 from client.screens.paint_window import PaintWindow
 
 
@@ -35,49 +23,12 @@ class PaintApp(App):
         return box
 
 
-class KeyboardManager:
-    def __init__(self, keyboard):
-        self._keyboard_shortcuts = {}
-        self.keycode_buffer = {}
-        self._keyboard = keyboard
-
-    def activate(self):
-        print("Binding keyboard")
-        self._keyboard.bind(on_key_down=self.on_key_down)
-        self._keyboard.bind(on_key_up=self.on_key_up)
-
-    def deactivate(self):
-        self._keyboard.unbind(on_key_down=self.on_key_down)
-        self._keyboard.unbind(on_key_up=self.on_key_up)
-
-    def is_key_down(self, keycode):
-        return keycode in self.keycode_buffer
-
-    def create_shortcut(self, shortcut, func):
-        if not isinstance(shortcut, tuple):
-            shortcut = (shortcut,)
-        self._keyboard_shortcuts[shortcut] = func
-
-    def on_key_down(self, keyboard, keycode, text, modifiers):
-        if keycode[1] in self.keycode_buffer:
-            return
-        self.keycode_buffer[keycode[1]] = keycode[0]
-
-    def on_key_up(self, keyboard, keycode):
-        for shortcut in self._keyboard_shortcuts.keys():
-            if keycode[1] in shortcut and set(
-                    shortcut).issubset(self.keycode_buffer):
-                self._keyboard_shortcuts[shortcut]()
-
-        self.keycode_buffer.pop(keycode[1])
-
-
 class DrawTool(MouseDrawnTool):
     def __init__(self, paint_window, **kwargs):
         super().__init__(**kwargs)
         self.paint_window = paint_window
         self.size = self.paint_window.size
-        self.keyboard = KeyboardManager(Window.request_keyboard(lambda: None, self))
+
         self.keyboard.create_shortcut(("lctrl", "z"), self.paint_window.undo)
         self.keyboard.create_shortcut(("lctrl", "y"), self.paint_window.redo)
         self.keyboard.create_shortcut("q", lambda: self.paint_window.set_visible(True))
